+++ boto3_refresh_session/methods/sts.py
@@ -90,14 +90,6 @@ class STSRefreshableSession(
         Botocore requires a successful refresh before continuing. If
         refresh fails in this window (in seconds), API calls may fail.
         Default is 10 minutes (600 seconds).
-    cache_clients : bool = True, optional
-        If ``True`` then clients created by this session will be cached and
-        reused for subsequent calls to :meth:`client()` with the same
-        parameter signatures. Due to the memory overhead of clients, the
-        default is ``True`` in order to protect system resources.
-    client_cache_max_size : int = 10, optional
-        The maximum number of clients to store in the client cache. Only
-        applicable if ``cache_clients`` is ``True``. Defaults to 10.
 
     Other Parameters
     ----------------
@@ -107,19 +99,22 @@ class STSRefreshableSession(
 
     Attributes
     ----------
-    client_cache : ClientCache
-        The client cache used to store and retrieve cached clients.
+    cache : SessionCache
+        The client and resource cache used to store and retrieve cached
+        clients.
     credentials : TemporaryCredentials
-        The current temporary AWS credentials.
+        The current temporary AWS security credentials.
 
     Methods
     -------
-    client(*args, **kwargs) -> boto3.client
-        Creates a Boto3 client for the specified service.
+    client(*args, eviction_policy: EvictionPolicy, max_size: int, **kwargs) -> BaseClient
+        Creates a low-level service client by name.
     get_identity() -> Identity
         Returns metadata about the current caller identity.
     refreshable_credentials() -> TemporaryCredentials
-        Returns the current temporary AWS credentials.
+        Returns the current temporary AWS security credentials.
+    resource(*args, eviction_policy: EvictionPolicy, max_size: int, **kwargs) -> ServiceResource
+        Creates a low-level service resource by name.
     whoami() -> Identity
         Alias for :meth:`get_identity`.
 
@@ -147,8 +142,8 @@ class STSRefreshableSession(
 
     .. note::
 
-        For additional details on client caching, refer to the
-        :ref:`client caching documentation <cachedocs>`.
+        For additional details on client and resource caching, refer to the
+        :ref:`caching documentation <cachedocs>`.
 
     Examples
     --------
@@ -181,7 +176,7 @@ class STSRefreshableSession(
     ...     mfa_token_provider=["ykman", "oath", "code", "<your-issuer>"],
     ...     mfa_token_provider_kwargs={"timeout": 45},  # custom timeout
     ... )
-    """
+    """  # noqa: E501
 
     def __init__(
         self,